+++ public/examples.me.js
@@ -1388,6 +1388,8 @@ var _manifest = _interopRequireDefault(require("../program/manifest"));
 
 var matrixEngine = _interopRequireWildcard(require("../index.js"));
 
+var _utility = require("../lib/utility.js");
+
 function _getRequireWildcardCache(nodeInterop) { if (typeof WeakMap !== "function") return null; var cacheBabelInterop = new WeakMap(); var cacheNodeInterop = new WeakMap(); return (_getRequireWildcardCache = function (nodeInterop) { return nodeInterop ? cacheNodeInterop : cacheBabelInterop; })(nodeInterop); }
 
 function _interopRequireWildcard(obj, nodeInterop) { if (!nodeInterop && obj && obj.__esModule) { return obj; } if (obj === null || typeof obj !== "object" && typeof obj !== "function") { return { default: obj }; } var cache = _getRequireWildcardCache(nodeInterop); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var key in obj) { if (key !== "default" && Object.prototype.hasOwnProperty.call(obj, key)) { var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null; if (desc && (desc.get || desc.set)) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } newObj.default = obj; if (cache) { cache.set(obj, newObj); } return newObj; }
@@ -1410,31 +1412,51 @@ var runThis = world => {
     source: ["res/images/complex_texture_1/diffuse.png"],
     mix_operation: "multiply"
   };
+  /**
+   * @description
+   * @note ONLY ANDROID DEVICES
+   * On simple CUBE works fine!
+   * For some reason on android devices it is not possible to use OBJ for video textures.
+   * Must be inspected - possible reasons:
+   * - Theres some limitation about vertex buffer data!
+   */
 
-  function onLoadObj(meshes) {
-    _manifest.default.meshes = meshes;
-    matrixEngine.objLoader.initMeshBuffers(world.GL.gl, _manifest.default.meshes.TV);
-    setTimeout(function () {
+  if ((0, _utility.isMobile)() == true) {
+    world.Add("cubeLightTex", 1, "TV", textuteImageSamplers);
+    _manifest.default.scene.TV.streamTextures = new ACCESS_CAMERA("webcam_beta");
+    (0, _utility.byId)('webcam_beta').style.display = 'block';
+    _manifest.default.scene.TV.rotation.rotationSpeed.y = 20; // Example who to switch between simple camera tex and cameraMixCanvas2d (videoImage)
+
+    _manifest.default.scene.TV.streamTextures.video = _manifest.default.scene.TV.streamTextures.videoImage;
+  } else {
+    function onLoadObj(meshes) {
+      _manifest.default.meshes = meshes;
+      matrixEngine.objLoader.initMeshBuffers(world.GL.gl, _manifest.default.meshes.TV);
       world.Add("obj", 1, "TV", textuteImageSamplers, _manifest.default.meshes.TV);
-      _manifest.default.scene.TV.position.y = 0;
-      _manifest.default.scene.TV.position.z = -4;
+      setTimeout(function () {
+        _manifest.default.scene.TV.position.y = 0;
+        _manifest.default.scene.TV.position.z = -4;
 
-      _manifest.default.scene.TV.rotation.rotateY(90);
+        _manifest.default.scene.TV.rotation.rotateY(90);
 
-      _manifest.default.scene.TV.LightsData.ambientLight.set(1, 1, 1);
+        _manifest.default.scene.TV.LightsData.ambientLight.set(1, 1, 1);
 
-      _manifest.default.scene.TV.streamTextures = new ACCESS_CAMERA("webcam_beta");
-    }, 1000);
+        (0, _utility.byId)('webcam_beta').style.display = 'block';
+        _manifest.default.scene.TV.streamTextures = new ACCESS_CAMERA("webcam_beta");
+      }, 1000);
+    }
+
+    matrixEngine.objLoader.downloadMeshes({
+      TV: "res/3d-objects/balltest2.obj"
+    }, onLoadObj);
   }
 
-  matrixEngine.objLoader.downloadMeshes({
-    TV: "res/3d-objects/balltest2.obj"
-  }, onLoadObj);
+  ;
 };
 
 exports.runThis = runThis;
 
-},{"../index.js":53,"../program/manifest":88}],17:[function(require,module,exports){
+},{"../index.js":53,"../lib/utility.js":79,"../program/manifest":88}],17:[function(require,module,exports){
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
@@ -2055,7 +2077,8 @@ var runThis = world => {
 
   window.addEventListener("contextmenu", e => {
     e.preventDefault();
-  }); // Override mouse up
+  });
+  matrixEngine.utility.createDomFPSController(); // Override mouse up
 
   _manifest.default.events.CALCULATE_TOUCH_UP_OR_MOUSE_UP = () => {
     console.log('TEST APP CLICK');
@@ -5859,35 +5882,61 @@ var webcamError = function (e) {
 exports.webcamError = webcamError;
 
 function SET_STREAM(video) {
-  if (navigator.getUserMedia) {
-    navigator.getUserMedia({
-      audio: true,
-      video: true
-    }, function (stream) {
-      try {
-        video.srcObject = stream;
-      } catch (error) {
-        video.src = window.URL.createObjectURL(stream);
+  var videoSrc = null;
+  navigator.mediaDevices.enumerateDevices().then(getDevices).then(getStream).catch(() => {
+    alert('ERR MEDIA');
+  });
+
+  function getDevices(deviceInfos) {
+    for (var i = 0; i !== deviceInfos.length; ++i) {
+      var deviceInfo = deviceInfos[i];
+
+      if (deviceInfo.kind === 'videoinput') {
+        videoSrc = deviceInfo.deviceId;
+        break;
       }
-    }, webcamError);
-  } else if (navigator.webkitGetUserMedia) {
-    navigator.webkitGetUserMedia({
-      audio: true,
-      video: true
-    }, function (stream) {
-      try {
-        video.srcObject = stream;
-      } catch (error) {
-        video.src = window.URL.createObjectURL(stream);
-      }
-    }, webcamError);
-  } else {
-    alert('webcam broken.');
+    }
+  }
+
+  function getStream() {
+    if (navigator.getUserMedia) {
+      navigator.getUserMedia({
+        audio: true,
+        video: {
+          deviceId: {
+            exact: videoSrc
+          },
+          facingMode: 'user'
+        }
+      }, function (stream) {
+        try {
+          console.log('stream1', stream);
+          video.srcObject = stream;
+          console.log('stream2', stream);
+        } catch (error) {
+          video.src = window.URL.createObjectURL(stream);
+        }
+      }, webcamError);
+    } else if (navigator.webkitGetUserMedia) {
+      navigator.webkitGetUserMedia({
+        audio: true,
+        video: true
+      }, function (stream) {
+        try {
+          video.srcObject = stream;
+        } catch (error) {
+          video.src = window.URL.createObjectURL(stream);
+        }
+      }, webcamError);
+    } else {
+      alert('webcam broken.');
+    }
   }
 }
 
 function ACCESS_CAMERA(htmlElement) {
   var ROOT = this;
+  console.log('?????????????????????????');
   ROOT.video = document.getElementById(htmlElement);
   SET_STREAM(ROOT.video);
   var DIV_CONTENT_STREAMS = document.getElementById('HOLDER_STREAMS');
@@ -8632,8 +8681,8 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
 
       if (object.streamTextures != null) {
         // video/webcam tex
-        // App.tools.loadVideoTexture('glVideoTexture', object.streamTextures.videoImage);
         if (object.streamTextures.video) {
+          // App.tools.loadVideoTexture('glVideoTexture', object.streamTextures.videoImage);
           _manifest.default.tools.loadVideoTexture('glVideoTexture', object.streamTextures.video);
         } else {
           _manifest.default.tools.loadVideoTexture('glVideoTexture', object.streamTextures.videoImage);
@@ -25298,6 +25347,8 @@ exports._glBlend = _glBlend;
 exports._DrawElements = _DrawElements;
 exports._glTexParameteri = _glTexParameteri;
 exports.gen2DTextFace = gen2DTextFace;
+exports.showDomFPSController = showDomFPSController;
+exports.createDomFPSController = createDomFPSController;
 exports.BiquadFilterType = exports.ENUMERATORS = exports.QueryString = exports.byId = exports.E = exports.scriptManager = exports.loadImage = exports.supportsTouch = exports.htmlHeader = exports.jsonHeaders = exports.HeaderTypes = void 0;
 
 var _manifest = _interopRequireDefault(require("../program/manifest"));
@@ -26040,6 +26091,146 @@ function gen2DTextFace(ctx, faceColor, textColor, text) {
   ctx.textBaseline = 'middle';
   ctx.fillStyle = textColor;
   ctx.fillText(text, width / 2, height / 2);
+}
+
+function showDomFPSController() {
+  byId('mobSpace').style.display = 'grid';
+  byId('mobRight').style.display = 'grid';
+  byId('mobLeft').style.display = 'grid';
+  byId('mobUp').style.display = 'grid';
+  byId('mobDown').style.display = 'grid';
+  byId('domAngleAxis').style.display = 'grid';
+} // Create DOM elements fort FPS template
+
+
+function createDomFPSController() {
+  var domSpace = document.createElement('div');
+  domSpace.id = 'mobSpace';
+  domSpace.classList.add('noselect');
+  domSpace.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 80%;
+      top: 80%;
+      width: 14%;
+      height: 4%;
+      background: rgba(255,255,255,0.2);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domSpace.innerText = `JUMP`;
+  domSpace.addEventListener('touchstart', e => {});
+  document.body.append(domSpace);
+  var domRight = document.createElement('div');
+  domRight.id = 'mobRight';
+  domRight.classList.add('noselect');
+  domRight.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 85%;
+      top: 90%;
+      width: 14%;
+      height: 4%;
+      background: rgba(255,255,255,0.2);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domRight.innerText = `RIGHT`;
+  domRight.addEventListener('touchstart', e => {
+    console.log('TEST RIGHT');
+  });
+  domRight.addEventListener('touchend', e => {});
+  document.body.append(domRight);
+  var domLeft = document.createElement('div');
+  domLeft.id = 'mobLeft';
+  domLeft.classList.add('noselect');
+  domLeft.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 70%;
+      top: 90%;
+      width: 14%;
+      height: 4%;
+      background: rgba(255,255,255,0.2);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domLeft.innerText = `LEFT`;
+  domLeft.addEventListener('touchstart', e => {
+    console.log('TEST domLeft');
+  });
+  domLeft.addEventListener('touchend', e => {});
+  document.body.append(domLeft);
+  var domUp = document.createElement('div');
+  domUp.id = 'mobUp';
+  domUp.classList.add('noselect');
+  domUp.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 78%;
+      top: 86%;
+      width: 14%;
+      height: 4%;
+      background: rgba(255,255,255,0.2);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domUp.innerText = `UP`;
+  domUp.addEventListener('touchstart', e => {
+    console.log('TEST domUp');
+  });
+  domUp.addEventListener('touchend', e => {
+    this.moveForward = false;
+  });
+  document.body.append(domUp);
+  var domDown = document.createElement('div');
+  domDown.id = 'mobDown';
+  domDown.classList.add('noselect');
+  domDown.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 78%;
+      top: 94%;
+      width: 14%;
+      height: 4%;
+      background: rgba(255,255,255,0.1);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domDown.innerText = `DOWN`;
+  domDown.addEventListener('touchstart', e => {});
+  domDown.addEventListener('touchend', e => {});
+  document.body.append(domDown);
+  var domAngleAxis = document.createElement('div');
+  domAngleAxis.id = 'domAngleAxis';
+  domAngleAxis.classList.add('noselect');
+  domAngleAxis.setAttribute('style', `
+      text-align: center;
+      display: none;
+      position:absolute;
+      left: 9%;
+      top: 82%;
+      width: 28%;
+      height: 28%;
+      background: rgba(255,255,255,0.1);
+      margin: auto;
+      align-items: center;
+      cursor: default;
+    `);
+  domAngleAxis.innerHTML = `    `;
+  domAngleAxis.addEventListener('touchstart', e => {});
+  document.body.append(domAngleAxis);
+  showDomFPSController();
 }
 
 },{"../program/manifest":88,"./events":55}],80:[function(require,module,exports){